requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
Pillow==8.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dotenv==0.16.0
pytz==2021.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
zope.interface==5.3.0
